% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{activities}
\alias{activities}
\title{Activities function}
\usage{
activities(activities = NULL)
}
\arguments{
\item{activities}{is the set of activities to retrieve}
}
\description{
This function queries the Activities API:
}
\keyword{activities}
