% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsites.R
\name{campsites}
\alias{campsites}
\title{Campsites function}
\usage{
campsites(lookup.type, lookup.args)
}
\arguments{
\item{lookup.type}{is the type of entity to use to query campsites}

\item{lookup.args}{is a vector of arguments used to query campsites}
}
\description{
This function queries the Campsites API:
}
\keyword{campsites}
