% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_version.R
\name{default.api.version}
\alias{default.api.version}
\title{Define package environment}
\description{
Sets default API version
}
\details{
Sets default API version. To change API version, use function set.api.version()
}
