% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilities.R
\name{facilities}
\alias{facilities}
\title{Facilities function}
\usage{
facilities(query.entity, query.id, data)
}
\arguments{
\item{data}{is a vector of facility-level data to obtain (e.g., campsites, events, addresses, etc.)}

\item{facilities}{is a set of facility IDs to query; if null, retrieve all facility IDs}

\item{org}{is a set of organization IDs for which to retrieve all facilities}

\item{rec.area}{is a set of RecArea IDs for which to retrieve all facilities}
}
\description{
This function queries the Facilities API:
}
\keyword{facilities}
