% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.geometry.R
\name{process.geometry}
\alias{process.geometry}
\title{Process geometry function}
\usage{
process.geometry(data)
}
\arguments{
\item{data}{is the relevant data from parsed API response}
}
\description{
This function process geometry returns from the FSF APIs:
}
\keyword{process.geometry}
