% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridb_query.R
\name{ridb.query}
\alias{ridb.query}
\title{RIDB query function}
\usage{
ridb.query(api, arg)
}
\arguments{
\item{api}{is one of RIDB APIs}

\item{arg}{is a query argument}
}
\description{
This function constructs an RIDB API request:
}
\keyword{ridb.query}
